Moved OrderContent struct to Content/OrderContent.swift file
